# Avoid expensive Throwable.fillInStackTrace call on each batch execution