When I profiled my application, I noticed suspicious calls to `Throwable.fillInStackTrace`.
```
	  at java.lang.Throwable.fillInStackTrace(Throwable.java:798)
	  - locked <0x2ce7> (a java.sql.SQLException)
	  at java.lang.Throwable.<init>(Throwable.java:256)
	  at java.lang.Exception.<init>(Exception.java:55)
	  at java.sql.SQLException.<init>(SQLException.java:142)
	  at org.h2.jdbc.JdbcPreparedStatement.executeBatch(JdbcPreparedStatement.java:1273)
	  at com.mchange.v2.c3p0.impl.NewProxyPreparedStatement.executeBatch(NewProxyPreparedStatement.java:2544)
	  at org.springframework.jdbc.core.JdbcTemplate.lambda$batchUpdate$5(JdbcTemplate.java:1143)
	  at org.springframework.jdbc.core.JdbcTemplate$$Lambda$724/0x000001e65b77aea8.doInPreparedStatement(Unknown Source:-1)
	  at org.springframework.jdbc.core.JdbcTemplate.execute(JdbcTemplate.java:651)
	  at org.springframework.jdbc.core.JdbcTemplate.execute(JdbcTemplate.java:691)
	  at org.springframework.jdbc.core.JdbcTemplate.batchUpdate(JdbcTemplate.java:1127)
```
![изображение](https://github.com/h2database/h2database/assets/741251/222e9b66-8502-4c6e-9ad4-4e42d63ff015)

I propose to use `ArrayList` instead of `SQLException` as _storage_ for batch exceptions.