@katzyn 2024-02-03
Thank you for your contribution!

Did you run any actual tests with JMH? It looks like your code actually works slower in warmed up JVM.