@katzyn 2024-02-03
I tested another case when there were exceptions.

I have an idea how to speed up both cases, but I need to test it.