@turbanoff 2024-02-03
>I tested another case when there were exceptions.

I think we are more interesting in case, when there are no exceptions. It's what usually happen during normal application execution.