@katzyn 2024-02-04
Your fix is good enough for original problem with performance, but there is another one and its solution requires reimplementation of the whole batch execution.